requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
djangorestframework==3.17.1
et_xmlfile==2.0.0
gunicorn==25.3.0
openpyxl==3.1.5
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
